successfully registered one user, all other bugs of user/controller, multer/middleware, user/model, cloudinary/config and dotenv.config fixed.

+++ src/utils/cloudinary.js
@@ -1,5 +1,7 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv";
+dotenv.config();
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -20,12 +22,15 @@ const uploadOnCloudinary = async (localFilePath) => {
     });
 
     // file uploaded successfully
-    console.log("File uploaded successfully, url: ", response.url);
+    // console.log("File uploaded successfully, url: ", response.url);
 
+    // fs.unlink(localFilePath)
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath);
-    return null;
+    if (localFilePath != "") {
+      fs.unlinkSync(localFilePath);
+    }
+    return error;
   }
 };
